fix(correcao de .yaml e pom.xml) preciso corrigir MissaoController. Corrigi relacionamentos de bancos de dados no codigo de models

+++ src/main/java/Infnet/Assessment/dto/MetricasMissaoDTO/MissaoMetricaDTO.java
@@ -1,4 +1,6 @@
 package Infnet.Assessment.dto.MetricasMissaoDTO;
+import java.math.BigDecimal;
+
 import Infnet.Assessment.enums.NivelPerigoMissao;
 import Infnet.Assessment.enums.StatusMissao;
 
@@ -7,5 +9,5 @@ public record MissaoMetricaDTO(
     StatusMissao status,
     NivelPerigoMissao nivelPerigo,
     Long qtdParticipantes,
-    Double totalRecompensas
+    BigDecimal totalRecompensas
 ) {}
